requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
autograd==1.8.0
autoray==0.7.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==41.0.7
cssselect2==0.8.0
diastatic-malt==2.15.2
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
filelock==3.18.0
Flask==3.1.2
fonttools==4.60.0
fsspec==2025.5.1
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.12.0
jmespath==1.0.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.4.0
networkx==3.5
numpy==2.3.1
openai==2.8.1
packaging==25.0
passlib==1.7.4
PennyLane==0.41.1
PennyLane_Lightning==0.41.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.10.1
pydantic_core==2.14.1
pydyf==0.11.0
PyMySQL==1.1.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
reportlab==4.0.7
requests==2.32.4
rsa==4.9.1
rustworkx==0.16.0
s3transfer==0.9.0
scipy==1.16.0
scipy-openblas32==0.3.30.0.1
serpapi==0.1.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
termcolor==3.1.0
tinycss2==1.4.0
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
weasyprint==60.2
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
